Find & check permissions on Org of the Project
Currently Launcher does pre-checks to make sure the identity running the launcher have required permissions to run rad lab launcher or to deploy respective module.

This enhancement will make sure that we are not running into issue/errors in permission pre-check when the parent is a folder instead of org for any project used for RAD Lab Management or an existing project to deploy the module.

We are recursively identifying the Org node of any folder / nested folder.